support completions for bash fish and zsh
